Se agrego manejo de perfiles

+++ src/app/modules/shared/components/confirm/confirm.component.ts
@@ -1,5 +1,6 @@
 import { Component, Inject, OnInit } from '@angular/core';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
+import { PerfilService } from '../../services/perfil.service';
 import { PqrService } from '../../services/pqr.service';
 
 @Component({
@@ -11,7 +12,8 @@ export class ConfirmComponent implements OnInit {
 
   constructor(public dialogRef: MatDialogRef<ConfirmComponent>,
               @Inject(MAT_DIALOG_DATA) public data: any,
-              private PqrService: PqrService) { }
+              private PqrService: PqrService,
+              private PerfilService: PerfilService) { }
 
   ngOnInit(): void {
   }
@@ -21,15 +23,33 @@ export class ConfirmComponent implements OnInit {
   }
 
   delete(){
-    if(this.data != null) {
-      this.PqrService.deletePqr(this.data.idPqr)
-          .subscribe((data: any) => {
-            this.dialogRef.close(1);
-          }, (error) =>{
-            this.dialogRef.close(2);
-          })
-    }else{
-      this.dialogRef.close(2);
+    if(this.data.id == 1){
+      console.log("entro")
+      console.log(this.data.idPqr)
+      if(this.data != null) {
+        this.PqrService.deletePqr(this.data.idPqr)
+            .subscribe((data: any) => {
+              this.dialogRef.close(1);
+            }, (error) =>{
+              this.dialogRef.close(2);
+            })
+      }else{
+        this.dialogRef.close(2);
+      }
+    }
+    if(this.data.id == 2){
+      console.log("entro")
+      console.log(this.data.idUser)
+      if(this.data != null) {
+        this.PerfilService.deletePerfil(this.data.idUser)
+            .subscribe((data: any) => {
+              this.dialogRef.close(1);
+            }, (error) =>{
+              this.dialogRef.close(2);
+            })
+      }else{
+        this.dialogRef.close(2);
+      }
     }
   }
